Get it working

+++ commands.d/00-clean.sh
@@ -2,9 +2,13 @@
 
 fleetctl list-unit-files -no-legend -fields unit | xargs fleetctl destroy
 
-echo "Murdering docker"
-for CONTAINER in "$(docker ps --quiet)"; do
-  docker kill $CONTAINER > /dev/null
-done
+CONTAINERS=$(docker ps --quiet)
+
+if [ -n "$CONTAINERS" ]; then
+  echo "Murdering docker"
+  for CONTAINER in "${CONTAINERS}"; do
+    docker kill $CONTAINER > /dev/null
+  done
+fi
 
 etcdctl rm --recursive /vulcand &> /dev/null
